Systems and Last Known Time Reporting Alert
Useful to detect if a system(s) and Last Known Time Reporting Alert indicate that the network of host is offline. 
SecOps means that a user may be an Insider THREAT and pulled the host of network temp offline to add malware or in order to exfiltrate date from the host.